Reuse database Dialog
