+++ extensions/mssql/src/mssql.d.ts
@@ -945,84 +945,58 @@ declare module 'mssql' {
 		}
 	}
 
-	export interface DatabaseViewInfo extends ObjectViewInfo<Database> {
-		loginNames: string[];
-		collationNames: string[];
-		compatibilityLevels: string[];
-		containmentTypes: string[];
-		recoveryModels: string[];
+	export interface IObjectManagementService {
+		/**
+		 * Initialize the object view and return the information to render the view.
+		 * @param contextId The context id of the view, generated by the extension and will be used in subsequent save/script/dispose operations.
+		 * @param objectType The object type.
+		 * @param connectionUri The original connection's URI.
+		 * @param database The target database.
+		 * @param isNewObject Whether the view is for creating a new object.
+		 * @param parentUrn The parent object's URN.
+		 * @param objectUrn The object's URN.
+		 */
+		initializeView(contextId: string, objectType: ObjectManagement.NodeType, connectionUri: string, database: string, isNewObject: boolean, parentUrn: string, objectUrn: string): Thenable<ObjectManagement.ObjectViewInfo<ObjectManagement.SqlObject>>;
+		/**
+		 * Save an object.
+		 * @param contextId The object view's context id.
+		 * @param object The object to be saved.
+		 */
+		save(contextId: string, object: ObjectManagement.SqlObject): Thenable<void>;
+		/**
+		 * Script an object.
+		 * @param contextId The object view's context id.
+		 * @param object The object to be scripted.
+		 */
+		script(contextId: string, object: ObjectManagement.SqlObject): Thenable<string>;
+		/**
+		 * Dispose a view.
+		 * @param contextId The id of the view.
+		 */
+		disposeView(contextId: string): Thenable<void>;
+		/**
+		 * Rename an object.
+		 * @param connectionUri The URI of the server connection.
+		 * @param objectType The object type.
+		 * @param objectUrn SMO Urn of the object to be renamed. More information: https://learn.microsoft.com/sql/relational-databases/server-management-objects-smo/overview-smo
+		 * @param newName The new name of the object.
+		 */
+		rename(connectionUri: string, objectType: ObjectManagement.NodeType, objectUrn: string, newName: string): Thenable<void>;
+		/**
+		 * Drop an object.
+		 * @param connectionUri The URI of the server connection.
+		 * @param objectType The object type.
+		 * @param objectUrn SMO Urn of the object to be dropped. More information: https://learn.microsoft.com/sql/relational-databases/server-management-objects-smo/overview-smo
+		 */
+		drop(connectionUri: string, objectType: ObjectManagement.NodeType, objectUrn: string): Thenable<void>;
+		/**
+		 * Search for objects.
+		 * @param contextId The object view's context id.
+		 * @param objectTypes The object types to search for.
+		 * @param searchText Search text.
+		 * @param schema Schema to search in.
+		 */
+		search(contextId: string, objectTypes: string[], searchText?: string, schema?: string): Thenable<ObjectManagement.SearchResultItem[]>;
 	}
-
-	export interface DatabasePropertiesInfo extends SqlObject {
-		collationName?: string;
-		dateCreated?: string;
-		lastDatabaseBackup?: string;
-		lastDatabaseLogBackup?: string;
-		memoryAllocatedToMemoryOptimizedObjectsInMb?: string;
-		memoryUsedByMemoryOptimizedObjectsInMb?: string;
-		numberOfUsers?: string;
-		owner?: string;
-		sizeInMb?: string;
-		spaceAvailableInMb?: string;
-		status?: string;
-	}
-
-	export interface DatabasePropertiesViewInfo extends ObjectViewInfo<DatabasePropertiesInfo> {
-	}
-}
-
-export interface IObjectManagementService {
-	/**
-	 * Initialize the object view and return the information to render the view.
-	 * @param contextId The context id of the view, generated by the extension and will be used in subsequent save/script/dispose operations.
-	 * @param objectType The object type.
-	 * @param connectionUri The original connection's URI.
-	 * @param database The target database.
-	 * @param isNewObject Whether the view is for creating a new object.
-	 * @param parentUrn The parent object's URN.
-	 * @param objectUrn The object's URN.
-	 */
-	initializeView(contextId: string, objectType: ObjectManagement.NodeType, connectionUri: string, database: string, isNewObject: boolean, parentUrn: string, objectUrn: string, dialogType?: string): Thenable<ObjectManagement.ObjectViewInfo<ObjectManagement.SqlObject>>;
-	/**
-	 * Save an object.
-	 * @param contextId The object view's context id.
-	 * @param object The object to be saved.
-	 */
-	save(contextId: string, object: ObjectManagement.SqlObject): Thenable<void>;
-	/**
-	 * Script an object.
-	 * @param contextId The object view's context id.
-	 * @param object The object to be scripted.
-	 */
-	script(contextId: string, object: ObjectManagement.SqlObject): Thenable<string>;
-	/**
-	 * Dispose a view.
-	 * @param contextId The id of the view.
-	 */
-	disposeView(contextId: string): Thenable<void>;
-	/**
-	 * Rename an object.
-	 * @param connectionUri The URI of the server connection.
-	 * @param objectType The object type.
-	 * @param objectUrn SMO Urn of the object to be renamed. More information: https://learn.microsoft.com/sql/relational-databases/server-management-objects-smo/overview-smo
-	 * @param newName The new name of the object.
-	 */
-	rename(connectionUri: string, objectType: ObjectManagement.NodeType, objectUrn: string, newName: string): Thenable<void>;
-	/**
-	 * Drop an object.
-	 * @param connectionUri The URI of the server connection.
-	 * @param objectType The object type.
-	 * @param objectUrn SMO Urn of the object to be dropped. More information: https://learn.microsoft.com/sql/relational-databases/server-management-objects-smo/overview-smo
-	 */
-	drop(connectionUri: string, objectType: ObjectManagement.NodeType, objectUrn: string): Thenable<void>;
-	/**
-	 * Search for objects.
-	 * @param contextId The object view's context id.
-	 * @param objectTypes The object types to search for.
-	 * @param searchText Search text.
-	 * @param schema Schema to search in.
-	 */
-	search(contextId: string, objectTypes: string[], searchText?: string, schema?: string): Thenable<ObjectManagement.SearchResultItem[]>;
-}
 	// Object Management - End.
 }
