+++ extensions/mssql/src/objectManagement/ui/databaseDialog.ts
@@ -7,33 +7,72 @@ import * as azdata from 'azdata';
 import { ObjectManagementDialogBase, ObjectManagementDialogOptions } from './objectManagementDialogBase';
 import { IObjectManagementService } from 'mssql';
 import * as localizedConstants from '../localizedConstants';
-import { CreateDatabaseDocUrl } from '../constants';
+import { CreateDatabaseDocUrl, DatabasePropertiesDocUrl } from '../constants';
 import { Database, DatabaseViewInfo } from '../interfaces';
 
 export class DatabaseDialog extends ObjectManagementDialogBase<Database, DatabaseViewInfo> {
-	private _nameInput: azdata.InputBoxComponent;
+	// Horizontal Tabs
+	private generalTab: azdata.window.DialogTab;
+	private filesTab: azdata.window.DialogTab;
+
+	private nameInput: azdata.InputBoxComponent;
+	private backupSection: azdata.GroupContainer;
+	private lastDatabaseBackupInput: azdata.InputBoxComponent;
+	private lastDatabaseLogBackupInput: azdata.InputBoxComponent;
+	private databaseSection: azdata.GroupContainer;
+	private statusInput: azdata.InputBoxComponent;
+	private ownerInput: azdata.InputBoxComponent;
+	private dateCreatedInput: azdata.InputBoxComponent;
+	private sizeInput: azdata.InputBoxComponent;
+	private spaceAvailabeInput: azdata.InputBoxComponent;
+	private numberOfUsersInput: azdata.InputBoxComponent;
+	private memoryAllocatedInput: azdata.InputBoxComponent;
+	private memoryUsedInput: azdata.InputBoxComponent;
+	private maintenanceSection: azdata.GroupContainer;
+	private collationInput: azdata.InputBoxComponent;
 
 	constructor(objectManagementService: IObjectManagementService, options: ObjectManagementDialogOptions) {
 		super(objectManagementService, options);
+		if (!options.isNewObject) {
+			this.generalTab = azdata.window.createTab(localizedConstants.GeneralSectionHeader);
+			this.filesTab = azdata.window.createTab(localizedConstants.FilesHeaderText);
+		}
 	}
 
 	protected override get helpUrl(): string {
-		return CreateDatabaseDocUrl;
+		return this.options.isNewObject ? CreateDatabaseDocUrl : DatabasePropertiesDocUrl;
 	}
 
 	protected async initializeUI(): Promise<void> {
-		let generalSection = this.initializeGeneralSection();
-		let optionsSection = this.initializeOptionsSection();
-		this.formContainer.addItems([generalSection, optionsSection]);
+		if (this.options.isNewObject) {
+			let generalSection = this.initializeGeneralSection();
+			let optionsSection = this.initializeOptionsSection();
+			this.formContainer.addItems([generalSection, optionsSection]);
+		} else {
+			// Initilaize general Tab sections
+			this.initializeBackupSection();
+			this.initializeDatabaseSection();
+			this.initializeMaintenanceSection();
+
+			this.registerTab(this.generalTab, [this.backupSection, this.databaseSection, this.maintenanceSection]);
+
+			this.dialogObject.content = [this.generalTab, this.filesTab];
+
+			// We need to close the already opened dialog during handleObjectPropertiesDialogCommand,
+			// which is required for the horizontal dialog tabs, as we cannot refresh the exising dialog
+			azdata.window.closeDialog(this.dialogObject);
+			azdata.window.openDialog(this.dialogObject);
+		}
 	}
 
+	//#region Create Database
 	private initializeGeneralSection(): azdata.GroupContainer {
 		let containers: azdata.Component[] = [];
-		this._nameInput = this.createInputBox(localizedConstants.NameText, async () => {
-			this.objectInfo.name = this._nameInput.value;
+		this.nameInput = this.createInputBox(localizedConstants.NameText, async () => {
+			this.objectInfo.name = this.nameInput.value;
 			await this.runValidation(false);
 		});
-		containers.push(this.createLabelInputContainer(localizedConstants.NameText, this._nameInput));
+		containers.push(this.createLabelInputContainer(localizedConstants.NameText, this.nameInput));
 
 		if (this.viewInfo.loginNames?.length > 0) {
 			this.objectInfo.owner = this.viewInfo.loginNames[0];
@@ -82,4 +121,68 @@ export class DatabaseDialog extends ObjectManagementDialogBase<Database, Databas
 
 		return this.createGroup(localizedConstants.OptionsSectionHeader, containers, true, true);
 	}
+	//#endregion
+
+	//#region Database Properties - General Tab
+	private initializeBackupSection(): void {
+		this.lastDatabaseBackupInput = this.createInputBox(localizedConstants.LastDatabaseBackupText, async () => { }, this.objectInfo.lastDatabaseBackup, this.options.isNewObject);
+		const lastDatabaseBackupContainer = this.createLabelInputContainer(localizedConstants.LastDatabaseBackupText, this.lastDatabaseBackupInput);
+
+		this.lastDatabaseLogBackupInput = this.createInputBox(localizedConstants.LastDatabaseLogBackupText, async () => { }, this.objectInfo.lastDatabaseLogBackup, this.options.isNewObject);
+		const lastDatabaseLogBackupContainer = this.createLabelInputContainer(localizedConstants.LastDatabaseLogBackupText, this.lastDatabaseLogBackupInput);
+
+		this.backupSection = this.createGroup(localizedConstants.BackupSectionHeader, [
+			lastDatabaseBackupContainer,
+			lastDatabaseLogBackupContainer
+		], false);
+	}
+
+	private initializeDatabaseSection(): void {
+		this.nameInput = this.createInputBox(localizedConstants.NamePropertyText, async () => { }, this.objectInfo.name, this.options.isNewObject);
+		const nameContainer = this.createLabelInputContainer(localizedConstants.NamePropertyText, this.nameInput);
+
+		this.statusInput = this.createInputBox(localizedConstants.StatusText, async () => { }, this.objectInfo.status, this.options.isNewObject);
+		const statusContainer = this.createLabelInputContainer(localizedConstants.StatusText, this.statusInput);
+
+		this.ownerInput = this.createInputBox(localizedConstants.OwnerPropertyText, async () => { }, this.objectInfo.owner, this.options.isNewObject);
+		const ownerContainer = this.createLabelInputContainer(localizedConstants.OwnerPropertyText, this.ownerInput);
+
+		this.dateCreatedInput = this.createInputBox(localizedConstants.DateCreatedText, async () => { }, this.objectInfo.dateCreated, this.options.isNewObject);
+		const dateCreatedContainer = this.createLabelInputContainer(localizedConstants.DateCreatedText, this.dateCreatedInput);
+
+		this.sizeInput = this.createInputBox(localizedConstants.SizeText, async () => { }, this.objectInfo.sizeInMb, this.options.isNewObject);
+		const sizeContainer = this.createLabelInputContainer(localizedConstants.SizeText, this.sizeInput);
+
+		this.spaceAvailabeInput = this.createInputBox(localizedConstants.SpaceAvailableText, async () => { }, this.objectInfo.spaceAvailableInMb, this.options.isNewObject);
+		const spaceAvailabeContainer = this.createLabelInputContainer(localizedConstants.SpaceAvailableText, this.spaceAvailabeInput);
+
+		this.numberOfUsersInput = this.createInputBox(localizedConstants.NumberOfUsersText, async () => { }, this.objectInfo.numberOfUsers, this.options.isNewObject);
+		const numberOfUsersContainer = this.createLabelInputContainer(localizedConstants.NumberOfUsersText, this.numberOfUsersInput);
+
+		this.memoryAllocatedInput = this.createInputBox(localizedConstants.MemoryAllocatedText, async () => { }, this.objectInfo.memoryAllocatedToMemoryOptimizedObjectsInMb, this.options.isNewObject);
+		const memoryAllocatedContainer = this.createLabelInputContainer(localizedConstants.MemoryAllocatedText, this.memoryAllocatedInput);
+
+		this.memoryUsedInput = this.createInputBox(localizedConstants.MemoryUsedText, async () => { }, this.objectInfo.memoryUsedByMemoryOptimizedObjectsInMb, this.options.isNewObject);
+		const memoryUsedContainer = this.createLabelInputContainer(localizedConstants.MemoryUsedText, this.memoryUsedInput);
+
+		this.databaseSection = this.createGroup(localizedConstants.DatabaseSectionHeader, [
+			nameContainer,
+			statusContainer,
+			ownerContainer,
+			dateCreatedContainer,
+			sizeContainer,
+			spaceAvailabeContainer,
+			numberOfUsersContainer,
+			memoryAllocatedContainer,
+			memoryUsedContainer
+		], false);
+	}
+
+	private initializeMaintenanceSection(): void {
+		this.collationInput = this.createInputBox(localizedConstants.CollationText, async () => { }, this.objectInfo.collationName, this.options.isNewObject);
+		const collationContainer = this.createLabelInputContainer(localizedConstants.CollationText, this.collationInput);
+
+		this.maintenanceSection = this.createGroup(localizedConstants.MaintenanceSectionHeader, [collationContainer], false);
+	}
+	//#endregion
 }
